Interviews now display employee picture
Added a nice profile picture frame for the employees. The implementation for the image is a little hacky, I am going to the '''database''' (json file) and finding the record where the filename matches and loading this image. So long as there are no duplicate file names this shouldn't be an issue. I couldn't get error images to work. I tried some jquery but it wouldn't work, something to look into for the future maybe?