Implemented hover states images
Updated the site image source to use formalImageUrl & informalImageUrl instead of imageUrl. Added hover state to interview image, formalImageUrl is displayed by default and imformalImageUrl is displayed on hover. The implementation could be prettier (i'd prefer it to be 2 <imgs> instead of hiding a div). But it works for what we need.